52_I2C_HDC1008.pm: fix "PERL WARNING: Use of uninitialized value"
This fixes "PERL WARNING: Use of uninitialized value in concatenation
(.) or string at ./FHEM/52_I2C_HDC1008.pm line 180". The {type} element
seems to be completely unused in the rest of the code, no idea what the
intention of the original author was when adding it here.

Thanks to joco for reporting! (Forum: #115543)